+++ obsidian/tests/param_configs.py
@@ -1,6 +1,9 @@
+"""Preset parameter configurations for unit testing"""
+
 from obsidian.parameters import Param_Continuous, Param_Ordinal, Param_Categorical, \
     Param_Observational, Param_Discrete_Numeric, ParamSpace
 
+# Set up ap master list of parameter spaces for testing
 params = [
         Param_Continuous('Parameter 1', 0, 10),
         Param_Continuous('Parameter 2', -20, 0),
@@ -15,14 +18,25 @@ params = [
         Param_Ordinal('Parameter 11', ['N'])
     ]
 
+# Subset some default selections
 default = [params[i] for i in [0, 1, 2, 6]]  # 2 continuous, 1 static, 1 categorical
+X_sp_default = ParamSpace(params=default)
+
+# Numeric
 cont_small = [params[i] for i in [0, 1, 2]]  # continuous including edge cases
 numeric = [params[i] for i in [0, 1, 2, 3, 4, 5]]  # numeric including edge cases
+X_sp_cont_small = ParamSpace(params=cont_small)
+X_sp_numeric = ParamSpace(params=numeric)
 
+# Nominal
 cat_small = [params[i] for i in [6, 7, 8]]  # categorical including edge cases
 disc_small = [params[i] for i in [6, 9]]  # 1 categorical, 1 ordinal
 disc_large = [params[i] for i in [6, 7, 8, 9, 10]]  # discrete including edge cases
+X_sp_cat_small = ParamSpace(params=cat_small)
+X_sp_disc_small = ParamSpace(params=disc_small)
+X_sp_disc_large = ParamSpace(params=disc_large)
 
+# Set up a range of continuous parameters
 params_cont_large = [
         Param_Continuous('Parameter 1', 0, 10),
         Param_Continuous('Parameter 2', 0, 10),
@@ -37,15 +51,8 @@ params_cont_large = [
         Param_Continuous('Parameter 11', 0, 10),
         Param_Continuous('Parameter 12', 0, 10),
 ]
-
-X_sp_default = ParamSpace(params=default)
-X_sp_cont_small = ParamSpace(params=cont_small)
 X_sp_cont_large = ParamSpace(params=params_cont_large)
-X_sp_numeric = ParamSpace(params=numeric)
-X_sp_cat_small = ParamSpace(params=cat_small)
-X_sp_disc_small = ParamSpace(params=disc_small)
-X_sp_disc_large = ParamSpace(params=disc_large)
-
-test_X_space = [X_sp_default, X_sp_cont_small, X_sp_numeric, X_sp_cat_small, X_sp_disc_small, X_sp_disc_large]
-
 X_sp_cont_ndims = [ParamSpace(params_cont_large[:i]) for i in range(len(params_cont_large))]
+
+# Wrap everything for iteration during testing
+test_X_space = [X_sp_default, X_sp_cont_small, X_sp_numeric, X_sp_cat_small, X_sp_disc_small, X_sp_disc_large]